+++ lib/compony/components/edit.rb
@@ -53,10 +53,8 @@ module Compony
             end
             hsh? local_form_comp.schema_wrapper_key_for(local_data), &local_form_comp.schema_block_for(local_data, local_controller)
           end
-          schema.validate!(controller.request.params)
-
-          # TODO: Why are we not saving the validated params?
-          attrs_to_assign = controller.request.params[form_comp.schema_wrapper_key_for(@data)]
+          validated_params = schema.validate!(controller.request.params)
+          attrs_to_assign = validated_params[form_comp.schema_wrapper_key_for(@data)]
           @data.assign_attributes(attrs_to_assign) if attrs_to_assign
         end